+++ Cargo.toml
@@ -26,6 +26,14 @@ async-io = "1.1.9"
 easy-parallel = "3.1.0"
 num_cpus = "1.13.0"
 scopeguard="1.1"
+criterion = "0.3"
 
 [profile.release]
-debug=1
+# incremental = true
+# debug = 0 # Set this to 1 or 2 to get more useful backtraces in debugger.
+# lto = true
+# codegen-units = 1
+
+[[bench]]
+name = "my_benchmark"
+harness = false
